requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
blessed==1.20.0
boto3==1.34.79
botocore==1.34.79
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.0.4
django-admin==2.0.2
django-excel-response2==3.0.6
django-extensions==3.2.3
django-six==1.0.5
django-storages==1.14.2
djangorestframework==3.15.1
excel-base==1.0.4
idna==3.6
isoweek==1.3.3
jmespath==1.0.1
pathspec==0.10.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
screen==1.0.1
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
TimeConvert==3.0.13
typing_extensions==4.11.0
tzlocal==5.2
urllib3==1.26.18
wcwidth==0.1.9
xlwt==1.3.0
